refactor: move UnitModal and CurriculumUnitDetails to CurriculumComponents
