+++ src/components/CurriculumComponents/CurriculumUnitDetails/CurriculumUnitDetails.test.tsx
