+++ src/components/CurriculumComponents/CurriculumUnitDetails/index.tsx
